address put-ttl problem
Dynamodb collects expired records within 48 hours. When a TTL is less than that time, this library protects against returning the record by checking the ttl (as per the best practice advised in the docs). This fix provides an 'expired: true' attribute on the err object which indicates that a key was not found because it has expired but has not been garbage collected. This is necessary in the case where using put() with 'notExists:true'.